AACGroupContainer.java - Fixed some variable and its related codes which is still using int type for item ID. - Rectified problem related to group deletion.
ActionGroup.java
- Fixed some variable and its related codes which is still using int type for item ID.
- Added expand_item_vector() by overriding its superclass ActionItem. It additionally adds its subtree items to the vector.

ActionItem.java
- Fixed some variable and its related codes which is still using int type for item ID.
- Added expand_item_vector(). This method expands given vector by adding an item. And if no vector is given, makes one and proceeds with it.

ActionMacro.java
- Fixed some variable and its related codes which is still using int type for item ID.

ActionWord.java
- Fixed some variable and its related codes which is still using int type for item ID.

MainActivity.java
- Renamed R.ids and xml files related to item selection mode.

+++ app/src/main/java/cwnuchrome/aac_cwnu_it_2015_1/ActionMacro.java
@@ -69,7 +69,7 @@ public class ActionMacro extends ActionMultiWord {
         return result;
     }
 
-    protected void addToRemovalList(Context context, AACGroupContainer.RemovalListBundle list, int id) {
+    protected void addToRemovalList(Context context, AACGroupContainer.RemovalListBundle list, long id) {
         list.add(ActionMain.item.ID_Macro, id);
     }
